+++ src/app/app.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 // import { ModalModule } from './sharedComponents/modal';
 
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { SidebarComponent } from './coreComponents/sidebar/sidebar.component';
@@ -10,8 +10,8 @@ import { NavBarComponent } from './sharedComponents/nav-bar/nav-bar.component';
 import { RecipesDisplayComponent } from './sharedComponents/recipes-display/recipes-display.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { FavoritesDisplayComponent } from './favorites-display/favorites-display.component';
-import { ModalComponent } from './sharedComponents/modal/modal.component';
-import { LoginModalComponent } from './coreComponents/modals/login-modal/login-modal.component';
+import { ModalComponent } from './modals/modal/modal.component';
+import { LoginModalComponent } from './modals/login-modal/login-modal.component';
 
 @NgModule({
   declarations: [
@@ -22,13 +22,13 @@ import { LoginModalComponent } from './coreComponents/modals/login-modal/login-m
     DashboardComponent,
     FavoritesDisplayComponent,
     ModalComponent,
-    LoginModalComponent
+    LoginModalComponent,
   ],
   imports: [
     FormsModule,
     BrowserModule,
     AppRoutingModule,
-    FormsModule
+    ReactiveFormsModule,
   ],
   providers: [],
   bootstrap: [AppComponent]
