--- src/app/coreComponents/modals/login-modal/login-modal.component.html
@@ -1,1 +1,0 @@
-<p>login-modal works!</p>
