--- src/app/coreComponents/modals/login-modal/login-modal.component.ts
@@ -1,25 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-import { ModalService } from '../../../modal.service';
-
-@Component({
-  selector: 'app-login-modal',
-  templateUrl: './login-modal.component.html',
-  styleUrls: ['./login-modal.component.scss']
-})
-export class LoginModalComponent implements OnInit {
-
-  constructor(private modalService: ModalService) { }
-
-  ngOnInit(): void {
-  }
-
-  openModal(id: string) {
-    this.modalService.open(id);
-  }
-
-  closeModal(id: string) {
-      this.modalService.close(id);
-  }
-
-}
