+++ src/app/modals/login-modal/login-modal.component.html
@@ -1,0 +1,21 @@
+<div>Log in to Fridge Forager!</div>
+<form [formGroup]="loginForm" (ngSubmit)="onSubmit()">
+  <div class="card">
+    <label class="card-title" for="username">
+      Username
+    </label>  
+    <div class="card-input-container">
+      <input class="card-input" type="text" formControlName="username" id="username">
+    </div>
+    <label class="card-title" for="password">
+      Password
+    </label>  
+    <div class="card-input-container">
+      <input class="card-input" type="text" formControlName="password" id="password">
+    </div>
+    <button class="confirm-button"type="submit" >
+      log in
+    </button>
+    </div>
+</form>
+<div>Register</div>
