+++ src/app/modals/modal/index.ts
@@ -1,3 +1,2 @@
 
-// export * from './modal.module';
 export * from '../../modal.service';
