+++ src/app/sharedComponents/nav-bar/nav-bar.component.html
@@ -1,4 +1,5 @@
-<div class='navbar'>Fridge Forager
+<div class='navbar'>
+  Fridge Forager
   <input 
   type="image" 
   [src]="favoriteIcon" 
@@ -9,8 +10,11 @@
   type="image" 
   [src]="userIcon" 
   class="nav-button"
-  (click)="openModal('login-modal')"  
+  (click)="openModal('login-modal', $event)"  
 />
+<modal id="login-modal">
+  <app-login-modal></app-login-modal>
+</modal>
 </div>
